requirements.txt
Django~=4.2.5
djangorestframework~=3.14.0
asgiref==3.7.2
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
